add restriction for other people's contacts

+++ src/services/auth.js
@@ -28,12 +28,13 @@ export const loginUser = async (payload) => {
   return newSession;
 };
 
-export const refreshUserSession = async ({ sessionId, refreshToken }) => {
+export const refreshUserSession = async (sessionId, refreshToken) => {
   const session = await SessionsCollection.findOne({
     _id: sessionId,
     refreshToken,
   });
   if (!session) throw createHttpError(401, 'Session not found');
+  if (session.refreshToken !== refreshToken) throw createHttpError(401, 'Refresh token is invalid');
   if (session.refreshTokenValidUntil < Date.now()) throw createHttpError(401, 'Session token expired');
 
   await SessionsCollection.deleteOne({ _id: session._id });
@@ -47,16 +48,14 @@ export const logoutUser = async (sessionId) => {
 };
 
 const createSession = async (userId) => {
-  // const accessTokenTTL = 15 * 60 * 1000; //FIFTEEN_MINUTES
-  // const refreshTokenTTL = 24 * 60 * 60 * 1000; //ONE_DAY
-  const accessTokenTTL = 2 * 60 * 1000; //two_MINUTES
-  const refreshTokenTTL = 3 * 60 * 1000; //three_MINUTES
+  const accessTokenTTL = 15 * 60 * 1000; //FIFTEEN_MINUTES
+  const refreshTokenTTL = 24 * 60 * 60 * 1000; //ONE_DAY
 
   const accessToken = randomBytes(30).toString('base64');
   const refreshToken = randomBytes(30).toString('base64');
 
   return await SessionsCollection.create({
-    userId: userId,
+    userId,
     accessToken,
     refreshToken,
     accessTokenValidUntil: new Date(Date.now() + accessTokenTTL),
